+++ qt/formatter.h
@@ -30,8 +30,8 @@ class Formatter: public QObject
 
     public:
 
-        static QString memToString( double bytes );
-        static QString sizeToString( double bytes );
+        static QString memToString( uint64_t bytes );
+        static QString sizeToString( uint64_t bytes );
         static QString speedToString( const Speed& speed );
         static QString percentToString( double x );
         static QString ratioToString( double ratio );
